IPTVTuningData: fix a broken cast to QApplication
Casting to a QApplication fails when run from mythackend since it
only creates a QCoreAppliction when started.

In this case we don't care if the instance returned is a QApplication
we only care if we dont have a QCoreApplication which is only the case
when this class is used by the unit tests.