Update transports with NIT delivery system descriptors
When the list of transports is extended with information found
in the delivery system descriptors in the NIT then do the scan
always with that information.
Previously, when there was already a transport at the frequency
specified in the NIT then the information in the NIT was
discarded and the old information in the transport was used.
This causes problems when the existing information is not correct.

Refs #13472