Add modulation system parameter in DVB-C "Full Scan (Tuned)"
Fixes an error message about undefined modulation system
when doing a DVB-C "Full Scan (Tuned)".
In that case the default modulation system DVB-C/A was used
which is probably correct for almost everybody but nevertheless
it is now possible to start a scan with the other DVB-C
modulation systems.
Also minor cosmetic changes such as more consistent naming
and ordering in the "Full Scan (Tuned)" option menu's
for DVB-S/S2, DVB-T/T2 and DVB-C.