Modulation system for DVB-T2 transports
Set the modulation system for DVB-T2 transport stream multiplexes to DVB-T2
if there is a T2 terrestial_delivery system descriptor present in the
NIT for that transport.
The HDHomeRun DVB-T/T2 tuners do automatically tune to whatever the
modulation system is and hence all transports are marked as DVB-T
after a channel scan with the HDHomeRun.
This is a problem when the same video source is also used with
other DVB-T/T2 tuners for whom this does make a difference
such as the MyGica T230 USB tuner.

The difference with the previous, reverted, implementation is that
now is searched explictly for the T2 delivery system descriptor instead
of for any extension descriptor. This makes a difference on systems
that actually have more extension descriptors than only the T2 descriptor.