Revert DVB-T bandwidth descriptor update
Change the string representation of the DVB-T bandwidth back to
only the number and add the "k" suffix in the toString of the descriptor.
This fixes adding of DVB-T transports from the NIT which
was broken after commit 1818e555f2d1b4d5b8193b5eeeb27f8e53b9f5f3.