Transport Editor updates
Show the DVB-C parameter edit page for HDHomeRun tuners when they can do DVB-C.
Previously the ATSC parameter edit page was shown.
Show the DVB-T2 parameter edit page for HDHomeRun tuners when they can do other DVB.
Previously the ATSC parameter edit page was shown.
Fixed a problem where sometimes the Transport Editor did not show any
transports when there were only a small number of transports in the video source.
FIxed a problem where the modulation system was not saved in the transport when a
DVB-S (not a DVB-S2) tuner and the transports came from the NIT transport loop.