Add support for the Service Relocated Descriptor
Add the Service Relocated Descriptor as described in
the A038 Blue Book page 115.
This descriptor is used when a service is moved from one
transponder to another and in that time the service does
appear in both transports.
Such a service appeared twice in the channel list.
This is now fixed; the service in the old location is removed
when the "Remove duplicates" scan option is selected.