tidy: Fix "redundant string initialization" warnings.
Having no initializer for a string, specifying a default initializer
for as string (aka {}), and specifying that a string be explicitly
initialized to a null string (aka {""}), all have the same result of
creating an empty string.  The clang-tidy program complains about the
last of these three initialization forms, so change the code to use
the default initializer so that it matches the style of the rest of
the initializations in that file.

https://clang.llvm.org/extra/clang-tidy/checks/readability-redundant-string-init.html