Don't double include object files in test code.
Mark a couple of functions as public (so their symbols are exported)
so that object files don't have to be explicitly include in the link
stage.  This prevents double including the objects, once explicitly
and once in the libmythtv library, which causes a weird double free on
only one platform.