tidy: Fix a memory leak in mythtv-setup.
This program allocates wither a QCoreApplication object or a
QApplication object using the "new" operator.  Clang-tidy flags this
as a memory leak as there is no corresponding "delete" operator.  (Its
there, just hidden elsewhere.)  This can't be converted to RAII
because the application object would get cleaned up too soon, so move
the explicit delete into the main function instead of having it
elsewhere.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-cplusplus.NewDeleteLeaks.html