tidy: Fix a memory leak in mythscreenwizard.
This program allocates the QApplication object using the "new"
operator.  Clang-tidy flags this as a memory leak as there is no
corresponding "delete" operator.  (Its there, just hidden elsewhere.)
Convert this to RAII so that the delete happens automatically in the
stack cleanup and clang-tidy understands that the object is deleted.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-cplusplus.NewDeleteLeaks.html