tidy: Fix a potential memory leaks in recordeingprofile.cpp.
The VideoCompressionSettings constructor has a memory leak when the
dynamic_res variable is set to false.  The code unconditionally
allocates three data structures, and uses all of them when dynamic_res
it true, but only uses one of them when dynamic_res is false.  Add
code to the false case to delete the two unused structure.

Use unique pointers to fix a couple of potential memory leaks flagged
by clang-tidy.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-cplusplus.NewDeleteLeaks.html