Remove conditional code checks for Qt < 5.9.
Qt 5.9 is required by Centos 7, SuSe 15.1, and Ubuntu 18.04.

NOTE: Debian 9 (Stretch) will no longer compile MythTV after this
commit.

The project has moved to the C++17 standard, and Debian 9 has a number
of gaps in its support of that standard.  Given that Debian 9 will be
End of Life two months after the next release of MythTV, it makes
sense to go ahead and drop support now and not spend time trying to
work around issues on a platform that won't exist after the next
release.