Strip quotes off of command line arguments.
With this change, the command line '-O "xyzzy"="plugh"' will be
treated the same as '-O xyzzy=plugh'.  They will both create an
override with a key of xyzzy that has a value of plugh.  Before this
change the former would create an override with a key named "xyzzy"
(quotes included) with a value of "plugh" (quotes included).

This applies to all command line arguments in the form "x=y".

This provides the same behavior as the override settings file, where
quoting has always been stripped.