tidy: Fix potential order-of-initialization problems.
A static value in mpegtables.cpp currently depends upon a static
variable initialized in tspacket.cpp, but there is no guarantee of the
order of initialization of the object files.  Convert this value (and
a couple of others) from a "static const" to a "static constexpr" so
that the constant value can be put into the header file and avoid any
order of initialization issues.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-interfaces-global-init.html