Update XMLTV loglevel in programdata
Change the loglevel in programdata to be the same as
EIT updates for the equivalent changes (i.e. debug).
This allows a run of mythfilldatabase with the options
of --verbose general,xmltv with the default loglevel
of info to not blather on about the expected case of
adding, deleting, updating programs (i.e. making the
usage easier to review), but just include the XMLTV
logging that likely matters for the average use case.

Last of four patches. Fixes #13633.

Signed-off-by: David Hampton <mythtv@love2code.net>
(cherry picked from commit 6e61aa988fe1cbc260798d5be53c496f66b04f4f)