Have unittests.sh set LD_LIBRARY_PATH on FreeBSD.
This is needed because of the way various systems handle interactions
between the RPATH specifications and the "--as-needed" linker flag.

RPATH specifies the runtime path to use when loading libraries.

The "--as-needed" linker flag affects how dependency information
is written into a program executable.  Without the "--as-needed"
linker flag, the linker will write dependency information into the
executable for every library listed on the command line.  With the
"--as-needed" flag the linker will only write dependency information
into the executable for libraries that are directly called from the
program.

On Linux and OSX, the RPATH specifications in the executable program
are used for loading all libraries.  If a library is not listed in the
executable but is listed as a dependency for a loaded library, it is
still the RPATH specifications from the program that are used for
loading the library.

On FreeBSD however, the RPATH specification in the executable is only
used for loading libraries *that* *are* *specified* *in* *the*
*executable*.  If one library depends upon another library, and that
library isn't listed in the executable because the "--as-needed" flag
was specified, then it is the RPATH values in the first library are
used to load the second library.  Since none of our libraries are
compiled with embedded RPATH specifications, the dependent library
cannot be found if isn't in the default system library locations.
FreeBSD does honor the LD_LIBRARY_PATH variable, so setting that
allows the libraries to be found even when the "--as-needed" linker
flag is used.