Discard PAT for wrong multiplex
When waiting for the PAT for a specific transport stream multiplex
discard PATs for other multiplexes.
After a new tuning request there can be still a number of packets received
from the previous multiplex and these packets can contain a PAT.
This fixes incorrect "Program not found in PAT" and "Rescan your transports"
messages that can sometimes be seen even immediately after a succesful scan.

Refs #13472