Default AVC/H264 parser in DTVRecorder
In DTVRecorder, after inspecting the packets, the AVC/H264 or the HEVC/H265 parser
is installed as needed. This works fine for recordings but this fails for
Live TV when the first channel is an MPEG2/SD channel and then a channel
change is done to a H264/HD channel.
In the latter case, there are packets processed without a parser being installed
and this then creates a segfault.
Fixed by installing the AVC/H264 parser in the constructor of DTVRecorder.