Switch the AlbumArt::addImage parameter from a reference to a pointer.
Marking the AlbumArt constructor as explicit pointed out that all the
callers to the addImage() function were passing a pointer where the
function expected a reference, and the compiler was implicitly
converting between the two.  Rewrite this function to take a pointer,
and thus eliminate all the implicit conversions from pointer to
reference. Add a couple of "const" keywords.

https://clang.llvm.org/extra/clang-tidy/checks/google-explicit-constructor.html