Utilize Schedules Direct metadata if grabber makes it available
Schedules Direct provides some additional metadata that at least one
upstream XMLTV grabber now optionally makes available.  Use that
metadata if it is provided by the grabber (typically the grabber will
need to be told to provide the additional information via a grabber
--configure).

In particular, with this patch, the original air date and a new
episode flag is available and can be useful to better schedule
recordings.

The (real soon now) next XMLTV release will include the new grabber,
and one can obtain the upstream master code now.

If the additional metadata information is not available (or not
configured in the grabber to be made available) the results are the
same as today.

Fixes #13655

Signed-off-by: David Engel <dengel@mythtv.org>