Handle tuning live TV before EIT data is collected.
To play Live TV, MythTV needs to create a recording and start playing
it. However the data needed to create the recording (title, category
and category type) hasn't yet been found so these fields are null and
cause the SQL queries to abort.  Check for these null query fields and
substitute "Unknown" or "Default" as appropriate.