Strip common suffixes from channel names when searching for icons.
Channel icon searching needs to be based on the actual channel name,
not one of the name variants that contains the picture definition
(e.g. "-SD" or "-HD") or the name of the technology used to transmit
the channel (e.g. "-DT").  Strip all these suffixes for better channel
matching.

Replace last QRegExp in file with a QRegularExpression.

Fix a debug message.