globalsettings: Add back a 'Paint engine' setting
- needed again to select Vulkan or OpenGL
- the Qt painter is never presented as an option but can be selected via
the command line again to try and get an app working (e.g. mythfrontend
-O PaintEngine=Qt) when there are OpenGL/Vulkan problems
- the setting is not displayed if there is no choice to be made (i.e.
vulkan support is not currently compiled in by default - so only OpenGL
will be available - so no setting is presented)
- the setting name is PaintEngine (the old ThemePainter setting will
likely have a value of 'auto' which is no longer supported) and avoids
hardcoded a workaround or performing a DB updateButton
- the code now builds a list of possible painters to try - with Qt as
the last option (unless overriden). This allows for better future
expansion.