No discontinuity for first TS packet of PID
Do not count the discontinuity from 0 to the continuity counter value of the first transport stream
packet of a PID as an error and consequently do also not give a "discontinuity detected" error message.
A perfect recording will now report a continuity_error_count of 0 instead of approximately the number of PIDs in the recording.