Remove test on PID before adding to list
Remove the test on the PID value not equal to 0 before adding the PID to the list of PIDS to listen to.
The comment indicates that the intention was to prevent adding the PID of program 0 to the list but the code does not match the comment.
The PID of program 0, when present in the PAT, is used to specify the PID of the NIT when it is not on the default PID.