No discontinuity for first TS packet of PID
Initialize all elements of m_continuityCounter to 0xff.
This is an essential part of the existing code that avoids giving continuity
error messages for the first transport stream packet received on a PID.
The initialization code is present in mythtv v30 but has been accidentally
removed moving forward to mythtv v31.