clazy: Fix using a const iterator with non-const functions.
The clazy "strict iterator" check pointed out a number of places where
a const iterator was declared, and then used with the results of
begin/end instead of constBegin/constEnd. If the container detaches
while running it in the method it will result in a crash.

These locations were manually running the list in reverse.  Use the Qt
supplied "const reverse" functions to handle this.