clazy: Convert const iterator with non-const functions to range-for. (plugins)
The clazy "strict iterator" check pointed out a number of places where
a const iterator was declared, and then used with the results of
begin/end/find instead of constBegin/constEnd/constFind. If the
container detaches while running it in the method it will result in a
crash.

Fix these by converting the loop to a c++11 range-for loop and use the
qAsConst function to prevent the Qt data structure from detaching.