clazy: Fix using a const iterator with non-const functions. (libs)
The clazy "strict iterator" check pointed out a number of places where
a const iterator was declared, and then used with the results of
begin/end/find instead of constBegin/constEnd/constFind. If the
container detaches while running it in the method it will result in a
crash.

Fix these by using the const versions of these functions.