Channelscan "Test Decryptability" fixes
When testing the decryptability of a program, process only the PIDs of that program.
Previously, also the PIDS of all previously tested programs in the same transport were processed.
Fix a locking order deadlock situation as described in ticket trac-11870 by releasing
the m_encryptionLock before calling the HandleEncryptionStatus of the listeners.
Improve the progress messages on the screen so that they fit in the available size
in the MythCenter-wide theme.

Refs #11870