Separate out CXXFLAGS from ECXXFLAGS in configure.
When the code in "configure" calls the add_cxxflags() function, the
new flag is currently added the value to ECXXFLAGS variable and not
the CXXFLAGS variable. That means that the compiler flag is never used
again during configure and is only passed on to the actual
compilations. This cause configure to fail when using
"--profile=hardened" because test compiles fail due to missing
arguments.

Fix this function to add to the CXXFLAGS variable, and add another
function for use by the "--extra-cxxflags" argument that adds to the
ECXXFLAGS variable. Update the qmake.pro files so that both of these
sets of flags are used in program compiles.