Fix audioconvert test failures when compiling X86 optimized code.
Tell the compiler what registers are clobbered by inline assembly
code. The compiler doesn't parse inline assembly except to do variable
substitution, so this is required to prevent the compiler from making
invalid assumptions about the lifetime of values stored in registers.

In this particular case, the compiler was trying to stash the value of
"f" value into one of the XMM registers, not realizing that register
was set to "NaN" (not a number) by the end of the inline assembly
code. This caused the loop at the end of the function to zero out the
final 0 to 15 values of the output array.