Fix avformatdecoder uninitialized variable warning.
This warning appeared when compiling with optimization and with
FORTIFY_SOURCE=2.

The CC608Decoder::GetServices function used to take an argument of
"bool[4]". Apparently this definition, while appearing to be pass by
value, was actually passing a pointer to the start of the array and
was therefore a pass by reference. Fix the new "std::array<bool,4>"
argument to also be pass by reference.