Time for preview max 10 minutes into the program
The time of the preview is 1/3 of the total program duration
limited to 10 minutes after the start of the program.
This limit is added to prevent accidentally revealing the
result of sport events in long recordings.

(cherry picked from commit b571d81175804f3e8cbc4c47e40232f1f76f0f96)
Signed-off-by: Klaas de Waal <klaas@kldo.nl>