Revert "Discard PAT for wrong multiplex"
Revert commit e05a8a841b8e88756ac0bb74322f9ed820862dc3 because of problems with active EIT in combination with Live TV.
When there is an active EIT scan running and then Live TV is started on the same capture card then the capture card
is tuned to the Live TV channel immediately without stopping the EIT scan. The PAT of the Live TV multiplex is then
ignored because it is different from the multiplex of the active EIT scan.