No table parsing on conditional access PIDs
Do not do table parsing on PIDs that carry conditional access information.
This data consists of entitlement messages etc that is completely private and proprietary.
Giving this data to the table parser leads to lots of error messages, for example:
2020-09-17 20:58:13.079216 I  PESPacket: Failed CRC check 0x28ef0f56 != 0x5ee48db2 for StreamID = 0xc8
2020-09-17 20:58:13.079225 E  MPEGStream[3](0x7fb9bc541c28): Discarding broken PSIP packet
This happens when mythbackend is running with "-v siparser" verbosity while receiving a
transport stream that has conditional access PIDs.