KeyBindings/MythControls: Add ability to filter out 'modifier' contexts
- this does not yet change any behaviour (as I haven't figured out to
change the UI without compatibility issues).
- the underlying issue here is that 'Long press' key overrides are not a
key binding as such - but a modification to key behaviour - which can
then be bound to.
- as such they do not really belong the in the keybindings UI - not
least because if a long press is setup, the user must exit the
keybindings page to save it before it can be used as a binding; which is
not user friendly

Refs #221