Fix incorrect artwork urls returned from ttvdb grabber
When performing a manual search for metadata for a video, the artwork
fetch was failing: no thumbnail was shown and no artwork was being
associated with the video. Ticket 13518 mentioned strange urls containing
"banners/_cache//banners". This commit fixes the bad urls and seems to
restore the downloading of the artwork. I don't know what led to the need
for this.