cppcheck: Remove several redundant conditions.
All of these are in the form "!A || (A && B)" which is equivalent to
"!A || B".  If A is false, the first clause is true and the whole
condition is true.  The only time the second clause is evaluated is
when A is true, so to actually test for that in the second clause is
unnecessary.