Fix running a metadata grabber twice in manual mode
Commit e81c7fd added accuracy of retrieving metadata in automatic mode,
but leads to some inefficiency in the use of manual mode:

If the search for an exact match was not successful, the other grabber
is called twice and appends its result twice.

This commit combines the intention of commit 2f9424c for non-automatic
mode and the commit for automatic mode (2738b98).

(cherry picked from commit 6f85a40f370bf8e9b88b6ab0b98f6f8bebee4d8f)