MythMusic: remove all cddb/freedb support
The CDDB lookup service from MusicBrainz and others has long since been
dicontinued which was causing long delays trying to use the non existant
web service when trying to play or rip a CD.

In the long term the lookup should get a MusicBrainz discid for a CD and
then use the MusicBrainz API to get the full metedata.

NOTE: This is just a quick and dirty fix and so has only been compile tested.
The media monitor is not working on Ubuntu which uses UDisks2 which MythTV
does not support so I'm unable to play a CD to test further.