CAM setup for multirec
Remove duplicate service IDs when setting up the CAM. The CAM needs to be setup only once for each
service also if more than one program is being recorded from the same service. This solves the issue
of recordings having interrupts at the moment a second recording is started on the same channel.
This can happen with overlapping recordings on the same channel, when the recording starts before
the scheduled start time or the recording continues after the scheduled end time.