Automated metadata lookup: Return if no match found
When performing a lookup in automatic mode, do not present a list
of all similar matches. Instead, terminate that lookup with a
failure notification only.

If a metadata lookup is done on a fresh and large video library,
the amount of popped up selection lists might get very big, and one
loses the context to what filename the selection list is meant for.