Metadata Lookup: Add an experimental feature in automatic mode
In case that no exact matches have been found during automatic lookup,
return the head of the list proposed by the grabber.

This helps a lot, if a tv-series name is given by the file names
in English, but the lookup was done in a different language.

Anyway, it unconditionally sets the first presented item, even the
match was not exactly found by MythTV.

One needs to set the environment variable "EXPERIMENTAL_METADATA_GRAB"
prior to the execution of mythfrontend or mythmetadatalookup.

(cherry picked from commit 75cf2569817437f5282776193d4ef11f8261fa2a)