libmythtv: Add MythPreviewPlayer
- as for other 'helper' player classes, this implements code specific to
preview generation
- BD and DVD previews are not supported
- NOTE: there was an implementaion of DVD seeking for a screen grab
within MythDVDPlayer but there was no way this code could ever be
triggered as the preview generator only ever created a base MythPlayer
class. Likewise for the code that prevented preview generation for BDs.
It is possible to add this functionality for DVDs (and probably BDs) but
it is likely to be error prone

Refs #264