iEIT on one recording per transport stream
With multirec it is possible to make more than one recording from one tuner, recording
two or more programs from the same transport stream multiplex or making multiple simultaneous
recordings of the same program in case of overlapping recordings that extend beyond the
scheduled end time.
When recording a program the EIT pid 0x12, which is part of the transport stream but not
part of a specific program, is added to the list of pids for that program and then
the EIT tables are processed.
When recording a second program from the same transport stream it should not have
the EIT pid added as the EIT processing is already done by the first recording.

What happened was that the EIT pid was frequenly added but also removed from the second recording.
This led to duplication of EIT processing in the second recording.
This is a waste of computing resources but in itself it does not cause problems.
However, because the pid was also sometimes removed from the list of pids there were
gaps in the stream of transport stream packets and this led to
a very large number of error messages from AddPacketTS, like the following:
E  AddTSPacket[42]: Out of sync!!! Need to wait for next payloadStart PID: 0x12, continuity counter: 13 (expected 7)

These messages are only shown show with the "-v record" verbosity selected so they will usually never be noticed.
This bug is now fixed by starting for each recording with empty lists of EIT pids to be added and to be removed
instead of creating empty lists for the stream and never clearing them.