libmythtv: Start refactoring MythFrame
- new C++ class is MythVideoFrame
- this initial commit creates the new class
- MythVideoFrame::CopyFrame replaces the old copy method - which is a
simple, hopefully clean copying of frame metadata and the underlying
buffers. Frame formats and size must match.
- custom code for converting from NV12 to YV12 is removed. This would
never be triggered with the existing code. The only time we have NV12
frames is when returned from hardware decoders - in which case we either
use them directly (OpenGL/Vulkan) or they are converted via the GPU to
YV12 - but even then this last case is not currently supported - as GPU
accelerated commflagging etc is not enabled.
- old copy frames test suite is replaced with new implementation to test
the new CopyFrame method

Refs #261