Fix longstanding issue with Scheduler::getConflicting().
Somewhere along the line, probably in a refactoring or cleanup,
getConflicting() was broken by making it use FindNextConflict().  The
problem is that FindNextConflict() strictly checks inputs and when
getConflicting() is called, the inputs for non-recording programs have
already been removed.  This fix adds an ignoreinput parameter to be
used when FindNextConflict() is called from getConflicting().