MythVideoFrame: Convert to using MythVideoFrame constructors
- this removes the last of the old VideoFrame code
- this appears to be working in all cases - BUT - MythVideoFrame now
takes ownership of the (software) buffer. I think I've caught all cases
of the buffer being deleted elsewhere but there may be spots I've
missed. There shouldn't be cases of leaks - but there may be double
frees (which will be obvious)
- minor cleanup/rename of the Clear function
- remove some unused code in MythAVUtil
- generally speaking, there is now a move away from customising offsets,
pitches and alignment - as it should just not be needed.

Refs #261 #267