Rename Mythdialog QVariant override functions to have unique names.
This prevents unexpected conversions to QVariant, that can cause
strange compiler warnings or strange program behavior. For example, a
argument that is a pointer to a function can be implicitly converted
to "QVariant(true)", and then the compiler complains that the address
of the function will never be NULL. Worse yet, this produces an
executable instead of failing, and the executable has missing
callbacks and doesn't function properly.