Table header in two transport stream packets
The table handling checks if the first part of the table header containing the table length is completely in the payload start packet.
If this not the case then the following message is given:
2020-10-20 22:41:30.070349 E  MPEGStream[3](0x7fb258158158): Error: AFCOffset(4)+StartOfFieldPointer(181)>184, pes length & current cannot be queried
and the packets are discarded until the next payload start.
However, this restriction is not specified in 13181-1 and the table handling is now extended to support this.

Tested with satellite Astra-1 19.2E, transport 1002, service 5021,  NHK WORLD-JPN.