Skip adaptation field in table payload packets
It has been observed that when the table header extends into the second packet that
there can be a zero-length adaptation field in the second packet before the payload.
The presence of this adaptation field was ignored and the byte specifying the length
of the adaptation field was erroneously included in the table.
This was then detected when the CRC check failed, resulting in this type of messages:
2020-10-20 22:41:27.980516 I  PESPacket: Failed CRC check 0x430b0111 != 0x5e91ff11 for StreamID = 0x1
2020-10-20 22:41:27.980535 E  MPEGStream[3](0x7fb258158158): Discarding broken PSIP packet
Although having an adaptation table at this place does not appear to be useful there is nothing
in 13181-1 that makes this non-conforming and hence the table handling is now extended to skip past
all adaptation fields that may be encountered in the second and subsequent packets that are part of
the table being extracted.

Tested with satellite Astra-1 19.2E, transport 1002, service 5021,  NHK WORLD-JPN.