Additional fields in mythtv-setup Channel Editor
Add sorting on service ID, transport ID and video source in the
channel list of the mythtv-setup Channel Editor.
Add service ID, frequency and transport ID to the fields that
can be displayed in the Channel Editor page of mythtv-setup.
Note that the selection of which values are displayed is done
in the UI theme, file config-ui.xml. This file needs to be
adapted for each theme to actually display the values.