More descriptors and formatted log output
Renamed the existing Audio Stream Descriptor by adding AC-3 prefix.
Added access functions and formatted log output for the
Audio Stream Descriptor and the Data Stream Alignment Descriptor.
Improved output of the Video Stream Descriptor.
This is used in the debug log output of mythtv-setup and
in the "-v siparser" log output of mythbackend.