clazy: Fix "incorrect emit" warning.
From the clazy docs: For readability purposes you should always use
emit (or Q_EMIT) when calling a signal.  Clazy will warn if you forget
to use emit (or Q_EMIT) or if you use them on a non-signal.

Clazy apparently wants the emit keyword to be on the same line as the
signal name that is being emitted, but in this case that causes the
clang-tidy "misleading indentation" warning to trigger because it sees
the file after it has been pre-processed and the emit keyword has been
removed. The compromise to make both happy is to use NOLINT to quiesce
the tidy warning.