Initialization of transport stream packet continuity counters
Move the initialzation of the transport stream packet continuity counters in DTVRecorder from
the constructor to the ResetForNewFile function so that it is also done on a LiveTV channel change.
This fixes false continuity counter error messages that appear when you change channels
in LiveTV and you change back to a channel that was selected before.