clazy: Fix "returning void expression" warnings.
This check looks for void functions that contain the expression:

   return doStuff();

across one or more lines, where doStuff is a function that returns
void. This compiles cleanly because the function returns void, but
appears to be a mistake in the code. Clean these up to be a call to
the function followed by a bare "return" (when necessary).