Add readonly status for editable fields
Add a readonly attribute to editable fields.
This makes it possible to disable editing of fields that should not be modified
but these fields are now still presented correct and it is possible to navigate
to these fields and to view the associated help message.
Previously setting fields readonly was done by setting fields to disabled.
This has the drawback that the visual presentation is greyed-out and that
the associated help text is never visible.