Wayland: Add workaround for invisible windows...
- Qt 5.11.2 works, Qt 5.14.2 is broken. I'm not sure when it was broken
and if it has subsequently been fixed... but here's the explanation...
- on newer versions of Qt when using Wayland and the xdg-shell
integration (and its variants), our painter window's window
(MythPainterWindow is a widget - which has a window) does not receive an
expose event on startup.
- as a result, calls to MythPainterWindow::update (called from
MythMainWindow::animate) do not result in paintEvent calls being
delivered and we get a blank screen
- this is easily identifiable by checking whether the window is visible
and not exposed (note - we only start checking once we have received a
resizeEvent, which should imply MythPainterWindow itself is visible)
- use a QTimer to start checking - and once the issue appears to be
fixed, stop the timer (note - I'm not sure whether the check should
continue or maybe restarted following a visibility change?)
- the issue is avoidable by forcing Qt to use a non xdg-shell - but the
other shell variants are slowly being deprecated as the xdg-shell spec
is finalised.

Refs #266 (but this does not fix it)